+++ custom-addons/cristal_agent/models/__init__.py
@@ -8,6 +8,7 @@ from . import agent_combo_line
 from . import agent_cadence
 from . import res_partner
 from . import crm_lead
+from . import sale_order
 from . import product_template
 from . import whatsapp_message
 from . import mail_message